requirements.txt
django
djangorestframework
psycopg2
djangorestframework_simplejwt
django-filter
